refactor: [#34] CategoryService 내 Category 별 Dish 그룹핑 수정
