+++ backend/src/main/java/com/codesquad/sidedish/category/CategoryService.java
@@ -1,9 +1,10 @@
 package com.codesquad.sidedish.category;
 
+import com.codesquad.sidedish.category.domain.Category;
+import com.codesquad.sidedish.category.dto.CategoryResponse;
 import com.codesquad.sidedish.dish.DishRepository;
 import com.codesquad.sidedish.dish.domain.Dish;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,14 +20,8 @@ public class CategoryService {
         List<Category> categories = categoryRepository.findBySectionName(sectionName);
         List<Dish> dishes = dishRepository.findBySectionName(sectionName);
 
-        Map<Integer, List<Dish>> dishMap = dishes.stream()
-            .collect(Collectors.groupingBy(Dish::getCategoryId));
-
-        for (Category category : categories) {
-            category.setDishes(dishMap.get(category.getId()));
-        }
         return categories.stream()
-            .map(CategoryResponse::from)
+            .map(category -> CategoryResponse.of(category, dishes))
             .collect(Collectors.toList());
     }
 }
